StoresModule: Transferring Prestashop invoice mapping to Mcitys invoice schema to the service PrestashopAdapter
