+++ src/modules/stores/interfaces/stores.interface.ts
@@ -1,4 +1,5 @@
 import { IDateFilter } from '../../../common/dates/datefilter.schema.js';
+import { IMcitysInvoice } from '../../accounting/schemas/mcitys/invoice.schema.js';
 
 export abstract class StoreAdapter {
 	abstract uploadProduct(): void;
@@ -7,7 +8,7 @@ export abstract class StoreAdapter {
 
 	abstract getProductCategories(): void;
 	abstract getProductAttributes(): void;
-
+	abstract mapInvoices(invoices: unknown): Promise<IMcitysInvoice[]>;
 	abstract getOrderDetailsByInvoiceID(invoiceID: number): Promise<unknown>;
 
 	abstract getLastInvoices(): Promise<unknown>;
